Remove WC_Log_Levels and WC_Logger_Interface stubs; centralize logging in tests
- Deleted the WC_Log_Levels and WC_Logger_Interface stub files to streamline the codebase.
- Introduced a new stubs.php file to provide necessary stubs for testing, ensuring compatibility and reducing redundancy.
- Updated references in the Logger class to use fully qualified class names for WC_Log_Levels, enhancing clarity and consistency.
