--- packages/php/blueprint/tests/Unit/stubs/WC_Log_Levels.php
@@ -1,17 +1,0 @@
-<?php
-
-if ( ! class_exists( 'WC_Log_Levels' ) ) {
-	/**
-	 * WC Log Levels Class
-	 */
-	class WC_Log_Levels {
-		const EMERGENCY = 'emergency';
-		const ALERT     = 'alert';
-		const CRITICAL  = 'critical';
-		const ERROR     = 'error';
-		const WARNING   = 'warning';
-		const NOTICE    = 'notice';
-		const INFO      = 'info';
-		const DEBUG     = 'debug';
-	}
-}
